extend provider interface to support provider source ranges (#3)
This adds an extension to the provider interface which allows for
dynamic lookup of the source IPs used for the monitor checks. This makes
it possible to implement an admission controller which automatically
adds the monitor provider source IPs to the
`nginx.ingress.kubernetes.io/whitelist-source-range` annotation of an
ingress whenever an ingress monitor is enabled for it.

Right now the code additions are not in use. Usage will be implemented
in a followup PR to keep the amount of changes easy to review.